Use Ring sessions for user authentication
Once user logs in, the server hands the client a cookie with a
`ring-session` key. This key is associated with session data in the
`leaflike.middlewares/all-sessions` atom. When an action that requires
authorization(eg. view all of user's bookmarks) takes place, the
`wrap-authorized` middleware checks that the client sending the
request is authorized to perform that action.